+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 
 # Change Log
 
-## Upcoming Release
+## 1.0.3 - January 2023
 
 * Added the [New-AbApplicationConsent](https://github.com/automationbrew/autobrew-powershell/blob/main/docs/help/New-AbApplicationConsent.md) command to create a new consent for an Azure Active Directory application in the specified Azure Active Directory tenant
 * Added the [New-AbApplicationGrant](https://github.com/automationbrew/autobrew-powershell/blob/main/docs/help/New-AbApplicationGrant.md) command to create a new application grant object that is used as part of the request to consent for an Azure Active Directory to access resources in an Azure Active Directory tenant